Fix bug where laser pointer/spotlight would not disappear sometimes
This was because the "idle monitor" provided by the system is inhibited
in various circumstances, such as apps being fullscreen or media
playing. This is so that the system isn't considered idle for the
purposes of whether it should e.g. go to sleep, but isn't what we
want - we just need to know when the mouse hasn't been moved in a time.
So now we just manage an idle timeout ourselves that we reset whenever
the mouse moves.

Also fixed a name error and changed another variable to match naming
conventions